Fix recursion in the base loop.
I've decided that calling .run_forever() recursively is to be forbidden,
and .run() is to be used when you want to run the event loop multiple
times. The idea is that .run_forever() should be the "global" or
"application-level" run, and should only be done once per application,
where as run is able to be called arbitrarily.