Fix race condition during Handle creation.
This fixes a race condition leading to a memory leak. The source can be
executed even before the handle was added to the handlers set. In that
case, it is only added after being run and never removed.

Fix this problem by adding it to the handlers set before attaching the
source.