Update bogus test to be a lot more obvious, if not as "good".
The previous test was trying to replicate a race condition
deterministically, which as it turns out only works when you've
accidentally imported gi.repository.GdkX11, because it does things to
GLib's main loop that I don't quite understand.

The test was completely flawed though, because even with the
iteration(True) call removed, the test passed, so instead we're going
with a lame "verify the calls happened in this order" type test.

Fixes #15.