Implement set_child_watcher on the event loop policy.
This is used by the asynctest library, so people that want to use it
with gbulb will need this code. The monkeypatching is an unfortunate
necessity.

I've implemented our own version of get_child_watcher while I'm here
because it previously lazily initialised a SafeChildWatcher, and there
was no way to make it instead do a GLibChildWatcher.

Fixes #16.