Prevent the GTK event loop from reusing the default GLib main context for every instance
This caused “fun” migration of all scheduled callbacks between event loop
instances with equally “fun” to debug asyncio low-level exceptions when
using multi-threading with several concurrently running event loops.

It is completely unclear to the author of this change why the original change
causes this breakage was introduced in 2015 in commit fde03cde in the first
place. Perhaps the intention was merely to use the GLib default main context
on the default main loop instance? Or perhaps its supposed to be shared between
all instances of the main loop on the UI/main thread (some kind of recursion
workaround maybe)?